Update Bus Tracking dan Fix Menu Maps

+++ app/src/main/java/com/example/vikar/publicbustracking/model/StationModel.java
@@ -36,7 +36,7 @@ public class StationModel {
         return longitude;
     }
 
-    public void setLongitude(int longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
